General fixes (#39)

+++ lib/src/data/pickup.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 abstract class PickUp {
   final double componentSize = 100;
-  final randomAmount = Random().nextDouble() * 1000;
+  final randomAmount = (Random().nextDouble() * 1000).toInt();
 
   void move(double dt);
 }
